Add interface file and fix typing issue to display favorites
