+++ src/Components/FavJokeContainer/FavJokeCard/FavJokeCard.tsx
@@ -1,36 +1,14 @@
-import './FavJokeCard.scss'
-import React from 'react'
+import './FavJokeCard.scss';
 
-export interface Joke {
-  icon?: string;
-  id: string;
-  chuckJoke: string;
-  isFavorited: boolean
-}
-
-// export interface UserJoke {
-//   textInput?: string,
-//   id: string,
-//   isFavorited: boolean
-// }
-
-interface Props {
-  joke: Joke;
-  userJoke: Joke;
-  favorites: Joke[];
-  handleFavoriting(joke: Joke): void
-}
- 
-
-const FavJokecard = ({chuckJoke, icon, id} : {chuckJoke: string, icon?: string, id: string}) => {
+const FavJokecard = ({chuckJoke, icon, id, isFavorited} : {chuckJoke: string, icon: string, id: string, isFavorited: boolean}) => {
   return (
     <section>
       <article className='favorite'>
         <p>{chuckJoke}</p>
       </article>
-      <img src={icon} alt='Chuck Icon'></img>
+      <img src={icon ? icon : 'https://assets.chucknorris.host/img/avatar/chuck-norris.png' } alt='Chuck Icon'></img>
       <button className='star'>⭐️</button>
-      {/* {props.Joke.isFavorited || props.userJoke.isFavorited ? <button className='star' onclick={props.handleFavoriting}>⭐️</button> : <button className='star' onclick={props.handleFavoriting}>✩</button>}       */}
+      {/* {isFavorited ? <button className='star' onclick={props.handleFavoriting}>⭐️</button> : <button className='star' onclick={props.handleFavoriting}>✩</button>}       */}
     </section>
   )
 }
